requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pandas==2.2.2
pydantic==2.7.1
